pyproject.toml
[project]
name = "chatryt-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-postgres>=3.0.5",
    "openai>=2.29.0",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
